requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Swagger-UI==3.36.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
redis==4.1.0
Flask-Caching==1.10.1 